requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
channels
daphne
python-decouple
Pillow
requests
textblob
django-cors-headers
asgiref